Made CraftDetectEvent cancellable, removed HitBoxDetectEvent
