+++ modules/api/src/main/java/net/countercraft/movecraft/events/CraftDetectEvent.java
@@ -1,12 +1,15 @@
 package net.countercraft.movecraft.events;
 
 import net.countercraft.movecraft.craft.Craft;
+import org.bukkit.event.Cancellable;
 import org.bukkit.event.HandlerList;
 import org.jetbrains.annotations.NotNull;
 
-public class CraftDetectEvent extends CraftEvent {
+public class CraftDetectEvent extends CraftEvent implements Cancellable {
 
     private static final HandlerList HANDLERS = new HandlerList();
+    private boolean cancelled;
+    private String failMessage = "";
 
     public CraftDetectEvent(@NotNull Craft craft) {
         super(craft);
@@ -20,5 +23,23 @@ public class CraftDetectEvent extends CraftEvent {
     @SuppressWarnings("unused")
     public static HandlerList getHandlerList() {
         return HANDLERS;
+    }
+
+    @Override
+    public boolean isCancelled() {
+        return cancelled;
+    }
+
+    @Override
+    public void setCancelled(boolean cancelled) {
+        this.cancelled = cancelled;
+    }
+
+    public String getFailMessage() {
+        return failMessage;
+    }
+
+    public void setFailMessage(String failMessage) {
+        this.failMessage = failMessage;
     }
 }
